Initial work on 'freezing' the date-time column.
